+++ src/components/App/App.test.tsx
@@ -1,54 +1,12 @@
-import type { Character } from '../cardItem/CardItem';
 import { render, screen, waitFor } from '@testing-library/react';
 import App from './App';
 import '@testing-library/jest-dom';
 import { fetchCharacters } from '../../services/service';
 import { describe, vi, it, expect, beforeEach } from 'vitest';
 import userEvent from '@testing-library/user-event';
+import { mockCharacters } from '../../tests/mockData';
 
 vi.mock('./style.css', () => ({}));
-const mockCharacters: Character[] = [
-  {
-    id: 1,
-    name: 'Rick Sanchez',
-    status: 'Alive',
-    species: 'Human',
-    type: '',
-    gender: 'Male',
-    image: 'https://rickandmortyapi.com/api/character/avatar/1.jpeg',
-    origin: {
-      name: 'Earth',
-      url: 'https://rickandmortyapi.com/api/location/1',
-    },
-    location: {
-      name: 'Earth',
-      url: 'https://rickandmortyapi.com/api/location/20',
-    },
-    episode: ['https://rickandmortyapi.com/api/episode/1'],
-    url: 'https://rickandmortyapi.com/api/character/1',
-    created: '2017-11-04T18:48:46.250Z',
-  },
-  {
-    id: 2,
-    name: 'Morty Smith',
-    status: 'Alive',
-    species: 'Human',
-    type: '',
-    gender: 'Male',
-    image: 'https://rickandmortyapi.com/api/character/avatar/2.jpeg',
-    origin: {
-      name: 'Earth',
-      url: 'https://rickandmortyapi.com/api/location/1',
-    },
-    location: {
-      name: 'Earth',
-      url: 'https://rickandmortyapi.com/api/location/20',
-    },
-    episode: ['https://rickandmortyapi.com/api/episode/1'],
-    url: 'https://rickandmortyapi.com/api/character/2',
-    created: '2017-11-04T18:50:21.651Z',
-  },
-];
 
 vi.mock('../../services/service', () => ({
   fetchCharacters: vi.fn().mockImplementation(() =>
