+++ src/components/main/main.test.tsx
@@ -2,56 +2,13 @@ import { describe, it, vi, expect } from 'vitest';
 import { fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import Main from './main';
-import type { Character } from '../cardItem/CardItem';
+import { mockCharacters } from '../../tests/mockData';
 
 vi.mock('../cardList/CardList', () => ({
   default: vi.fn(() => <div data-testid="mock-card-list" />),
 }));
 
 describe('Main Component', () => {
-  const mockCharacters: Character[] = [
-    {
-      id: 1,
-      name: 'Rick Sanchez',
-      status: 'Alive',
-      species: 'Human',
-      type: '',
-      gender: 'Male',
-      image: 'https://rickandmortyapi.com/api/character/avatar/1.jpeg',
-      origin: {
-        name: 'Earth',
-        url: 'https://rickandmortyapi.com/api/location/1',
-      },
-      location: {
-        name: 'Earth',
-        url: 'https://rickandmortyapi.com/api/location/20',
-      },
-      episode: ['https://rickandmortyapi.com/api/episode/1'],
-      url: 'https://rickandmortyapi.com/api/character/1',
-      created: '2017-11-04T18:48:46.250Z',
-    },
-    {
-      id: 2,
-      name: 'Morty Smith',
-      status: 'Alive',
-      species: 'Human',
-      type: '',
-      gender: 'Male',
-      image: 'https://rickandmortyapi.com/api/character/avatar/2.jpeg',
-      origin: {
-        name: 'Earth',
-        url: 'https://rickandmortyapi.com/api/location/1',
-      },
-      location: {
-        name: 'Earth',
-        url: 'https://rickandmortyapi.com/api/location/20',
-      },
-      episode: ['https://rickandmortyapi.com/api/episode/1'],
-      url: 'https://rickandmortyapi.com/api/character/2',
-      created: '2017-11-04T18:50:21.651Z',
-    },
-  ];
-
   const defaultProps = {
     characters: mockCharacters,
     loading: false,
